combine storage implementations into storage.py, some cleanup
