--- snews/IStorage.py
@@ -1,12 +1,0 @@
-class IStorage(object):
-    def insert(self, time, neutrino_time, message):
-        pass
-
-    def getAllMessages(self):
-        pass
-
-    def cacheEmpty(self):
-        pass
-
-    def getMsgFromStrID(self, post_id):
-        pass
